+++ apps/public-docsite-v9-headless/.storybook/headless-docs-page.css
@@ -1,0 +1,31 @@
+/*
+  Loaded once from .storybook/preview.js. Drives the docs-page chrome around
+  the canvas card so HeadlessSourcePanel can portal its tabbed code panel into
+  the same bordered preview.
+*/
+
+.headless-docs-page .sbdocs-preview > *:not(.docs-story):not(.headless-source-portal) {
+  display: none !important;
+}
+
+.headless-docs-page .sbdocs-preview:has(> .headless-source-portal:not(:empty)) {
+  height: auto !important;
+}
+
+/*
+  Force the magenta accent for the "Show code" / "Open in Stackblitz" hover &
+  focus underlines. Storybook's ActionBar paints the underline via an inset
+  box-shadow driven by `theme.color.secondary`, and the
+  `@fluentui/react-storybook-addon-export-to-sandbox` styles hard-code a blue
+  underline on the Stackblitz button — both are overridden here so the canvas
+  action buttons match the rest of the headless docs accent.
+*/
+.headless-docs-page .sbdocs-preview .docblock-code-toggle:hover,
+.headless-docs-page .sbdocs-preview .docblock-code-toggle:focus,
+.headless-docs-page .sbdocs-preview .docblock-code-toggle.docblock-code-toggle--expanded,
+.headless-docs-page .docs-story .with-code-sandbox-button:hover,
+.headless-docs-page .docs-story .with-code-sandbox-button:focus {
+  outline: none !important;
+  box-shadow: #9b1f5a 0 -3px 0 0 inset !important;
+  color: #9b1f5a !important;
+}
