+++ packages/react-components/react-headless-components-preview/stories/.storybook/preview.js
@@ -7,11 +7,6 @@ import * as rootPreview from '../../../../../.storybook/preview';
 // storybook (built by `pr-website-deploy.yml`) renders identical stories.
 import '../theme/tokens.css';
 
-// Custom docs page that renders the "Show code" panel with TSX | CSS tabs.
-// See `stories/src/_helpers/HeadlessDocsPage.tsx` for the rationale (Storybook's
-// built-in Source block can't be made multi-language via MDX overrides).
-import { HeadlessDocsPage } from '../src/_helpers/HeadlessDocsPage';
-
 polyfillBodyAndObserve();
 
 /** @type {typeof rootPreview.decorators} */
@@ -20,10 +15,6 @@ export const decorators = [...rootPreview.decorators];
 /** @type {typeof rootPreview.parameters} */
 export const parameters = {
   ...rootPreview.parameters,
-  docs: {
-    ...(rootPreview.parameters && rootPreview.parameters.docs),
-    page: HeadlessDocsPage,
-  },
 };
 
 export const tags = ['autodocs'];
